[doors] door rework + button-controlled doors
