+++ Assets/Prefabs/guard/Guard.prefab
@@ -445,7 +445,7 @@ GameObject:
   - component: {fileID: 6166422522389729070}
   m_Layer: 7
   m_Name: Vision Cone
-  m_TagString: Untagged
+  m_TagString: GuardVisionCone
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
