+++ Assets/Prefabs/objects/ButtonDoor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 543d4aa098def024b807e8ff62f458d5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
