+++ Assets/Prefabs/objects/ClickDoor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 75a60e4d59de1354b95d639128adb6a5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
