+++ Assets/Prefabs/player/Player.prefab
@@ -76,7 +76,7 @@ SpriteRenderer:
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: 0
   m_SortingLayer: 0
-  m_SortingOrder: 0
+  m_SortingOrder: 1
   m_Sprite: {fileID: -2413806693520163455, guid: a86470a33a6bf42c4b3595704624658b, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
@@ -125,7 +125,7 @@ MonoBehaviour:
   moveSpeed: 10
   x_axis: 
   y_axis: 
-  collider: {fileID: -3522677471127249613}
+  circleCollider: {fileID: 0}
   rangeDisplayObject: {fileID: 7860588898707812371}
 --- !u!58 &-3522677471127249613
 CircleCollider2D:
