+++ Assets/Scripts/objects/ButtonDoorScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 217f82c5c09d06f459b04094ee8dd10b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
