+++ Assets/Scripts/objects/ButtonScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e07c78242380ed47bc51ab80954ba16
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
