+++ Assets/Scripts/objects/ClickDoorBehavior.cs
@@ -1,0 +1,26 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ClickDoorBehavior : DoorScript
+{
+    public GameObject spriteOutlineObject;
+
+    protected override void Update()
+    {
+        base.Update();
+
+        spriteOutlineObject.SetActive(GameState.Instance.ClickControllerScript.IsTargetObject(gameObject));
+    }
+    public void OnClick()
+    {
+        if (closed)
+        {
+            closed = false;
+        }
+        else
+        {
+            closed = true;
+        }
+    }
+}
