+++ Assets/Scripts/objects/ClickDoorBehavior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d39f93c09e958614d93c078ba104dce9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
