--- Assets/Scripts/objects/DoorBehavior.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 35657aaba85641f46a61eaf8df8f4290
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
