+++ Assets/Scripts/objects/DoorGuardColliderHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 10a7db5b082fbdb4db7a3c11e2656a64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
