+++ Assets/Scripts/objects/DoorScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 76bf2dbb8029edb439eff81ce57cdf93
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
